requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
gensim==4.3.3
nltk==3.9.1
joblib==1.4.2

fastapi==0.115.0
uvicorn[standard]==0.30.6

pytest==8.3.2
ruff==0.6.9

python-dotenv==1.0.1
httpx==0.27.2


